requirements.txt
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.12.3
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dalex==1.6.0
debugpy==1.8.12
decorator==5.2.1
Deprecated==1.2.18
exceptiongroup==1.2.2
executing==2.2.0
fonttools==4.56.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==6.0.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mysql-connector==2.2.9
mysql-connector-repackaged==0.3.1
narwhals==1.27.1
nest-asyncio==1.6.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyzmq==26.2.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.5.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.5.0
wcwidth==0.2.13
wrapt==1.17.3
